refactor: Simplify ContentPresenter by moving shared logic to shared file
