+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenter.crossruntime.cs
@@ -9,23 +9,7 @@ namespace Microsoft.UI.Xaml.Controls;
 
 partial class ContentPresenter
 {
-	private readonly BorderLayerRenderer _borderRenderer;
-
-	public ContentPresenter()
-	{
-		_borderRenderer = new BorderLayerRenderer(this);
-		InitializeContentPresenter();
-	}
-
-	private void SetUpdateTemplate() => UpdateContentTemplateRoot();
-
 	partial void RegisterContentTemplateRoot() => AddChild(ContentTemplateRoot);
 
 	partial void UnregisterContentTemplateRoot() => RemoveChild(ContentTemplateRoot);
-
-	private void UpdateCornerRadius(CornerRadius _) => UpdateBorder();
-
-	private void UpdateBorder() => _borderRenderer.Update();
-
-	partial void OnPaddingChangedPartial(Thickness oldValue, Thickness newValue) => UpdateBorder();
 }
