+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenterExtensions.cs
@@ -2,18 +2,14 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Microsoft.UI.Xaml.Controls
+namespace Microsoft.UI.Xaml.Controls;
+
+public static class ContentPresenterExtensions
 {
-	public static class ContentPresenterExtensions
-	{
-		public static DataTemplate ResolveContentTemplate(this ContentPresenter contentPresenter)
-		{
-			return DataTemplateHelper.ResolveTemplate(
-				contentPresenter?.ContentTemplate,
-				contentPresenter?.ContentTemplateSelector,
-				contentPresenter?.Content,
-				contentPresenter
-			);
-		}
-	}
+	public static DataTemplate ResolveContentTemplate(this ContentPresenter contentPresenter) =>
+		DataTemplateHelper.ResolveTemplate(
+			contentPresenter?.ContentTemplate,
+			contentPresenter?.ContentTemplateSelector,
+			contentPresenter?.Content,
+			contentPresenter);
 }
